+++ src/components/initial-loader.tsx
@@ -33,9 +33,9 @@ export const InitialLoader = () => {
   useEffect(() => {
     status &&
       status.length > 0 &&
-      handleStatusRedirect(status[0]['applicationStatus'], history)
+      handleStatusRedirect(status[0]['applicationStatus'], status[0]['applicationType'], history)
   }, [status])
 
-  return error ? <ErrorPage /> : <div className="initial-loader"></div>
+  return (error || !status?.length) ? <ErrorPage /> : <div className="initial-loader"></div>
 }
 export default withRouter(InitialLoader)